Revert the queryMap
