+++ packages/commerce-sdk-react/src/hooks/utils.ts
@@ -4,13 +4,25 @@
  * SPDX-License-Identifier: BSD-3-Clause
  * For full license text, see the LICENSE file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
-import {QueryClient} from '@tanstack/react-query'
+import {QueryClient, QueryKey} from '@tanstack/react-query'
 import {ApiClients, Argument, DataType} from './types'
 import {ShopperCustomersMutationType} from './ShopperCustomers'
 import {ShopperOrdersMutationType} from './ShopperOrders'
 import {ShopperBasketMutationType} from './ShopperBaskets'
 
-const isObject = (item: null) => typeof item === 'object' && !Array.isArray(item) && item !== null
+const isObject = (item: unknown) =>
+    typeof item === 'object' && !Array.isArray(item) && item !== null
+
+interface QueryKeyMap {
+    name: string
+    key: QueryKey
+}
+
+export interface CacheUpdateMatrixElement {
+    update?: Array<QueryKeyMap>
+    invalidate?: Array<QueryKeyMap>
+    remove?: Array<QueryKeyMap>
+}
 
 export interface QueryKeysMatrixElement {
     update?: Array<Array<string | unknown>>
@@ -28,6 +40,10 @@ type QueryKeysMatrix = {
     [key in CombinedMutationTypes]?: (data: any, param: any) => QueryKeysMatrixElement
 }
 
+type CacheUpdateMatrix = {
+    [key in CombinedMutationTypes]?: (data: any, param: any) => CacheUpdateMatrixElement
+}
+
 // TODO: Add more api clients as needed
 export type Client = ApiClients['shopperOrders'] &
     ApiClients['shopperCustomers'] &
@@ -36,37 +52,37 @@ export type Client = ApiClients['shopperOrders'] &
 export const updateCache = <Action extends CombinedMutationTypes>(
     queryClient: QueryClient,
     action: Action,
-    queryKeysMatrix: QueryKeysMatrix,
+    cacheUpdateMatrix: CacheUpdateMatrix,
     response: DataType<Client[Action]>,
     params: Argument<Client[Action]>
 ) => {
     // TODO: review this
-    const isMatchingKey = (cacheQuery: {queryKey: {[x: string]: any}}, queryKey: any[]) =>
+    const isMatchingKey = (cacheQuery: {queryKey: {[x: string]: any}}, queryKey: QueryKey) =>
         queryKey.every((item, index) =>
             isObject(item) && isObject(cacheQuery.queryKey[index])
                 ? Object.entries(cacheQuery.queryKey[index])
                       .sort()
                       .toString() ===
-                  Object.entries(item)
+                  Object.entries(item as Record<string, unknown>)
                       .sort()
                       .toString()
                 : item === cacheQuery.queryKey[index]
         )
 
     // STEP 1. Update data inside query cache for the matching queryKeys
-    queryKeysMatrix[action]?.(params, response)?.update?.map((queryKey: any) => {
+    cacheUpdateMatrix[action]?.(params, response)?.update?.map(({key: queryKey}) => {
         queryClient.setQueryData(queryKey, response)
     })
 
     // STEP 2. Invalidate cache entries with the matching queryKeys
-    queryKeysMatrix[action]?.(params, response)?.invalidate?.map((queryKey: any) => {
+    cacheUpdateMatrix[action]?.(params, response)?.invalidate?.map(({key: queryKey}) => {
         queryClient.invalidateQueries({
             predicate: (cacheQuery: any) => isMatchingKey(cacheQuery, queryKey)
         })
     })
 
     // STEP 3. Remove cache entries with the matching queryKeys
-    queryKeysMatrix[action]?.(params, response)?.remove?.map((queryKey: any) => {
+    cacheUpdateMatrix[action]?.(params, response)?.remove?.map(({key: queryKey}) => {
         queryClient.removeQueries({
             predicate: (cacheQuery: any) => isMatchingKey(cacheQuery, queryKey)
         })
